Make compute_cost_euro_millicents abstract on Executor
